Don't reset GOPATH when installing tools
I ... don't know why I did that, and even with modules the GOPATH is
used (as a cache, in GOPATH/pkg/mod). I think the main reason I reset it
is either because it used to work different (I originally wrote this in
vgo days) or maybe just simple ignorance on my part.

Seems to work well without resetting GOPATH anyway; just setting
GO111MODULE=on is enough to ensure modules are used, and setting GOBIN
ensures the correct bin path is used.

Fixes #17