Highlight cgo directives, only run goPackageComment when folding is enabled
I'm not sure why I always applied the folding pattern (it's been a few
years since I wrote it), but I can't find any good reason for it, and as
far as I can see it works well. Maybe some other change affected this?
Or maybe I was just stupid...

It's rather slow for larger files; and actually runs out of memory in
gtk.go in mattn/go-gtk.